requirements.txt
PyQt6 == 6.7.1
pillow == 10.4.0
